// packages/ui/.storybook/types.ts
export type Viewports =
	| 'iphone5'
	| 'iphone6'
	| 'iphone6p'
	| 'iphone8p'
	| 'iphonex'
	| 'iphonexr'
	| 'iphonexsmax'
	| 'iphonese2'
	| 'iphone12mini'
	| 'iphone12'
	| 'iphone12promax'
	| 'ipad'
	| 'ipad10p'
	| 'ipad12p'
	| 'galaxys5'
	| 'galaxys9'
	| 'nexus5x'
	| 'nexus6p'
	| 'pixel'
	| 'pixelxl'